refactor(flagr-ui): address PR review — merge constraint SCSS and placeholders
- Move constraint-table.scss into app.scss; drop separate @use in App.vue
- Fold property/value placeholder helpers into constraintOperatorUi.ts
- Remove constraintFieldPlaceholders.ts and constraint-table.scss
