--- browser/flagr-ui/src/styles/constraint-table.scss
@@ -1,116 +1,0 @@
-// Constraint grid: lives in global CSS so child SFC roots (display: contents) participate in the parent grid.
-.constraint-table {
-  display: grid;
-  grid-template-columns: 40px minmax(0, 1fr) minmax(0, 1.28fr) minmax(0, 1fr) auto;
-  column-gap: var(--space-2xs);
-  row-gap: var(--space-3xs);
-  align-items: center;
-  margin-top: var(--space-3xs);
-}
-
-.constraint-row {
-  display: contents;
-}
-
-.constraint-logic {
-  font-size: var(--font-size-micro);
-  font-weight: var(--font-weight-bold);
-  color: var(--el-text-color-secondary);
-  letter-spacing: var(--letter-spacing-wide);
-  text-transform: uppercase;
-  text-align: right;
-  justify-self: end;
-  align-self: center;
-  white-space: nowrap;
-}
-
-.constraint-logic--add {
-  color: var(--el-color-primary);
-}
-
-.constraint-cell {
-  min-width: 0;
-  align-self: center;
-}
-
-.constraint-control {
-  width: 100%;
-
-  .el-input__wrapper,
-  .el-select__wrapper {
-    border-radius: var(--radius-sm);
-  }
-}
-
-.constraint-actions {
-  display: flex;
-  gap: var(--space-3xs);
-  align-self: center;
-  justify-self: end;
-}
-
-.constraint-hint-cell {
-  grid-column: 2 / -1;
-  align-self: start;
-}
-
-.constraint-hint-line {
-  margin: 0 0 var(--space-3xs);
-}
-
-.constraint-add-caption {
-  grid-column: 1 / -1;
-  margin: 0;
-  font-size: var(--font-size-caption);
-  font-weight: var(--font-weight-semibold);
-  letter-spacing: var(--letter-spacing-ui);
-  color: var(--el-text-color-secondary);
-}
-
-.constraint-row--add {
-  opacity: 0.88;
-
-  .constraint-logic--add {
-    color: var(--el-text-color-placeholder);
-    font-size: var(--font-size-tab);
-    font-weight: 600;
-    line-height: 1;
-  }
-
-  .constraint-control .el-input__wrapper,
-  .constraint-control .el-select__wrapper {
-    background-color: var(--el-fill-color-lighter);
-  }
-}
-
-.constraint-add-divider {
-  grid-column: 1 / -1;
-  height: 0;
-  border-top: 1px dashed var(--el-border-color-lighter);
-  margin: var(--space-3xs) 0 var(--space-3xs);
-}
-
-.constraint-add-btn {
-  justify-self: end;
-  align-self: center;
-}
-
-@media (max-width: 720px) {
-  .constraint-table {
-    grid-template-columns: 36px 1fr;
-  }
-
-  .constraint-actions {
-    grid-column: 2;
-    justify-self: start;
-  }
-
-  .constraint-hint-cell {
-    grid-column: 2;
-  }
-
-  .constraint-add-btn {
-    grid-column: 2;
-    justify-self: start;
-  }
-}
